+++ lib/app_theme.dart
@@ -1,4 +1,10 @@
-library app_theme;
+import 'package:flutils/theme/material_theme.dart';
+import 'package:flutils/theme/text_theme.dart';
 
-export 'theme/material_theme.dart';
-export 'theme/text_theme.dart';
+class AppTheme {
+  /// Returns the [MaterialTheme].
+  static MaterialTheme get materialTheme => MaterialTheme();
+
+  /// Returns the [TextTheme].
+  static TextTheme get textTheme => TextTheme();
+}