+++ lib/bottom_sheets.dart
@@ -1,3 +1,4 @@
+/// This library is a wrapper for all the bottom sheets.
 library bottom_sheets;
 
 export 'bottom_sheets/input_bottom_sheet.dart';