+++ lib/dialogs.dart
@@ -1,3 +1,4 @@
+/// This library is a wrapper for all the dialogs.
 library dialogs;
 
 export 'dialogs/confirmation_dialog.dart';