+++ lib/theme/text_theme.dart
@@ -2,12 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 class TextTheme {
-  /// This constructor prevents instantiation and extension.
-  TextTheme._();
-
   /// White text with font sizes between 0 and 100
   /// Example: TextTheme.white[20]
-  static List<TextStyle> white = List<TextStyle>.generate(
+  List<TextStyle> white = List<TextStyle>.generate(
     100,
     (size) => GoogleFonts.poppins(
       color: Colors.white,
@@ -17,7 +14,7 @@ class TextTheme {
 
   /// White bold text with font sizes between 0 and 100
   /// Example: TextTheme.whiteBold[20]
-  static List<TextStyle> whiteBold = List<TextStyle>.generate(
+  List<TextStyle> whiteBold = List<TextStyle>.generate(
     100,
     (size) => GoogleFonts.poppins(
       color: Colors.white,
@@ -28,7 +25,7 @@ class TextTheme {
 
   /// Grey[300] text with font sizes between 0 and 100
   /// Example: TextTheme.grey[20]
-  static List<TextStyle> grey = List<TextStyle>.generate(
+  List<TextStyle> grey = List<TextStyle>.generate(
     100,
     (size) => GoogleFonts.poppins(
       color: Colors.grey[300],
@@ -38,7 +35,7 @@ class TextTheme {
 
   /// Grey[300] bold text with font sizes between 0 and 100
   /// Example: TextTheme.greyBold[20]
-  static List<TextStyle> greyBold = List<TextStyle>.generate(
+  List<TextStyle> greyBold = List<TextStyle>.generate(
     100,
     (size) => GoogleFonts.poppins(
       color: Colors.grey[300],